fixes broken symlinks
the version of rdoc that gets installed has changed, so the path needed to be updated.

What's more, this now ensures the target exists - to avoid creating links that don't actually point to a file. As a bonus, this will fail the next time the rdoc version changes, so we can fix accordingly.

Signed-off-by: Yorgos Saslis <yorgos.saslis@redhat.com>